requirements.txt
python-dotenv>=1.0
langchain>=0.2
langgraph>=0.0.38
langchain-openai>=0.1
langchain-community>=0.0.38
beautifulsoup4>=4.12
